requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
psycopg2-binary==2.9.1